Refactored in usecases testing package
