+++ src/test/java/usecases/gamesave/SaveLoadTest.java
@@ -1,4 +1,4 @@
-package usecases;
+package usecases.gamesave;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
