+++ src/test/java/usecases/sound/SoundInteractorTest.java
@@ -1,4 +1,4 @@
-package usecases;
+package usecases.sound;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
